model/requirements.txt
kaggle
tqdm
google-cloud-storage
PyPDF2
requests
numpy
sentence-transformers
pinecone-client
kaggle
python-dotenv
